\name{ProbeDeveloper-package}
\alias{ProbeDeveloper-package}
\alias{ProbeDeveloper}
\docType{package}
\title{
Develop Hybridization Probes
}
\description{
This tool can develop hybridization probes for target sequences based on melting temperature value calculated by R package 'TmCalculator' <https://CRAN.R-project.org/package=TmCalculator> and methods extended from Beliveau, B. J.,(2018) <doi:10.1073/pnas.1714530115>,  which can be used to capture specific target regions in fluorescence in situ hybridization and next generation sequence experiments.
}

\author{
Junhui Li

Maintainer: Junhui Li <junhuili@cau.edu.cn>
}
\references{
Beliveau B J, Kishi J Y, Nir G, et al. (2018). OligoMiner: A rapid, flexible environment for the design of genome-scale oligonucleotide in situ hybridization probes. bioRxiv.
}
