\name{samplefa}
\alias{samplefa}
\docType{data}
\title{
sample data for target sequence region with class 'DNAStringSet'
}
\description{
sample data read by function readDNAStringSet in R package 'Biostrings' from fasta format file, there are two target sequence resion in this data
}
\usage{data("samplefa")}
\format{
  Formal class 'DNAStringSet' [package "Biostrings"] with 5 slots
  \describe{
    sample data read by function readDNAStringSet in R package 'Biostrings' from fasta format file, which is from ncbiRefSeq database for Homo Sapiens with referece genome version hg19
  }
}
\examples{
data(samplefa)
}
\keyword{datasets}
